requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.63
langchain-openai==0.2.0
requests==2.32.5
python-dotenv==1.2.1
pydantic==2.12.4
streamlit==1.40.0
